content_gen/requirements.txt
python-fasthtml
claudette
starlette